requirements-uv.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile .\requirements.txt -o requirements-uv.txt --index-strategy unsafe-best-match --no-build-isolation
filelock==3.16.1
    # via torch
fsspec==2024.12.0
    # via torch
imageio==2.36.1
    # via -r ./requirements.txt
imageio-ffmpeg==0.5.1
    # via -r ./requirements.txt
jinja2==3.1.5
    # via torch
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.1
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.2.1
    # via
    #   imageio
    #   pandas
    #   pylance
    #   torchvision
packaging==24.2
    # via pillow-jxl-plugin
pandas==2.2.3
    # via -r ./requirements.txt
pillow==10.4.0
    # via
    #   imageio
    #   pillow-heif
    #   pillow-jxl-plugin
    #   torchvision
pillow-avif-plugin==1.4.6
    # via -r ./requirements.txt
pillow-heif==0.21.0
    # via -r ./requirements.txt
pillow-jxl-plugin==1.3.0
    # via -r ./requirements.txt
pyarrow==18.1.0
    # via
    #   -r ./requirements.txt
    #   pylance
pygments==2.18.0
    # via rich
pylance==0.20.0
    # via -r ./requirements.txt
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.2
    # via pandas
rich==13.9.4
    # via -r ./requirements.txt
setuptools==75.6.0
    # via imageio-ffmpeg
six==1.17.0
    # via python-dateutil
sympy==1.13.1
    # via torch
toml==0.10.2
    # via -r ./requirements.txt
torch==2.5.1+cu124
    # via
    #   -r ./requirements.txt
    #   torchvision
torchvision==0.20.1+cu124
    # via -r ./requirements.txt
typing-extensions==4.12.2
    # via
    #   rich
    #   torch
tzdata==2024.2
    # via pandas
